Adds energistic flare
Adds 'Flare', a new Operant Energistic ability allowing you to flash someone in melee range by targeting their eyes on harm intent.
The ability also functions like a standard flash, being able to disable both borgs and simple animals, but requires that your Psi-Ocular Luminescence is active to function at all.

The strength of the flash scales with Energistics rank:
- Operant is 3.
- Master is 4.5.
- Grandmaster is 6.
- Paramount is 7.5.
For comparison, a regular flash has a minimum of 1 and a maximum of 7.